Move image config building to frontend package
