+++ frontend/mariner2/handle_container.go
@@ -48,7 +48,7 @@ func handleContainer(ctx context.Context, client gwclient.Client, spec *dalec.Sp
 		return nil, nil, err
 	}
 
-	img, err := dalec.BuildImageConfig(ctx, client, spec, targetKey, marinerDistrolessRef)
+	img, err := frontend.BuildImageConfig(ctx, client, spec, targetKey, marinerDistrolessRef)
 	if err != nil {
 		return nil, nil, err
 	}
